Add hero name field

+++ contracts/game/HeroManager.sol
@@ -24,7 +24,7 @@ contract HeroManager is Ownable, Multicall, Randomness {
   uint256 public secondaryMultiplier = 8;
   uint256 public thirdMultiplier = 6;
 
-  uint256 public bonusExp = 30; // From Level 1, every battle win will give 30 exp to the hero. And as level goes up, this will be reduced. Level 1 -> 2: 30, Lv 2 -> 3: 29, ...., Lv 29 -> 30: 2
+  uint8 public bonusExp = 30; // From Level 1, every battle win will give 30 exp to the hero. And as level goes up, this will be reduced. Level 1 -> 2: 30, Lv 2 -> 3: 29, ...., Lv 29 -> 30: 2
 
   uint256 public rarityPowerBooster = 110;
 
@@ -39,7 +39,7 @@ contract HeroManager is Ownable, Multicall, Randomness {
     heroes[heroId] = hero;
   }
 
-  function levelUp(uint256 heroId, uint256 levels) public {
+  function levelUp(uint256 heroId, uint8 levels) public {
     require(nft.ownerOf(heroId) == msg.sender, "HeroManager: not a NFT owner");
 
     GameFi.Hero memory hero = heroes[heroId];
@@ -126,7 +126,7 @@ contract HeroManager is Ownable, Multicall, Randomness {
     return heroes[heroId].level;
   }
 
-  function heroBonusExp(uint256 heroId) public view returns (uint256) {
+  function heroBonusExp(uint256 heroId) public view returns (uint8) {
     return bonusExp - heroes[heroId].level + 1;
   }
 
@@ -137,7 +137,7 @@ contract HeroManager is Ownable, Multicall, Randomness {
     );
 
     if (heroes[heroId].level < 30) {
-      uint256 gainedExp = heroBonusExp(heroId);
+      uint8 gainedExp = heroBonusExp(heroId);
       heroes[heroId].experience += gainedExp;
       if (heroes[heroId].experience >= 100) {
         heroes[heroId].experience -= 100;
@@ -164,7 +164,7 @@ contract HeroManager is Ownable, Multicall, Randomness {
     );
 
     if (heroes[heroId].level > 1 || heroes[heroId].experience > 0) {
-      uint256 gainedExp = heroBonusExp(heroId) / 2;
+      uint8 gainedExp = heroBonusExp(heroId) / 2;
       heroes[heroId].experience -= gainedExp;
       if (heroes[heroId].experience < 0) {
         if (heroes[heroId].level == 1) {
@@ -216,7 +216,7 @@ contract HeroManager is Ownable, Multicall, Randomness {
     bonusLevelUpFee = value;
   }
 
-  function setBonusExp(uint256 value) external onlyOwner {
+  function setBonusExp(uint8 value) external onlyOwner {
     bonusExp = value;
   }
